+++ examples/auth-with-shared-root/components/App.js
@@ -44,4 +44,4 @@ const App = React.createClass({
 
 })
 
-export default App
+module.exports = App