+++ modules/__tests__/Link-test.js
@@ -516,4 +516,45 @@ describe('A <Link>', () => {
     })
   })
 
+  describe('when the "to" prop is unspecified', function () {
+    class App extends Component {
+      render() {
+        return (
+          <div>
+            <Link>Blank Link</Link>
+            <Link/>
+            <Link className="kitten-link">Kittens</Link>
+          </div>
+        )
+      }
+    }
+
+    it('returns an anchor tag without an href', function (done) {
+      render((
+        <Router history={createHistory('/')}>
+          <Route path="/" component={App} />
+        </Router>
+      ), node, function () {
+        const link1 = node.querySelectorAll('a')[0]
+        const link2 = node.querySelectorAll('a')[1]
+        const link3 = node.querySelectorAll('a')[2]
+        expect(link1.href).toEqual('')
+        expect(link2.href).toEqual('')
+        expect(link3.href).toEqual('')
+        done()
+      })
+    })
+
+    it('passes down other props', function (done) {
+      render((
+        <Router history={createHistory('/')}>
+          <Route path="/" component={App} />
+        </Router>
+      ), node, function () {
+        const link3 = node.querySelectorAll('a')[2]
+        expect(link3.className).toEqual('kitten-link')
+        done()
+      })
+    })
+  })
 })