+++ CHANGES.md
@@ -41,6 +41,40 @@
 [#3446]: https://github.com/reactjs/react-router/pull/3446
 
 
+## [v2.8.1]
+> Sep 13, 2016
+
+- **Bugfix:** Fix redirects that specify `query` ([#3808])
+
+[v2.8.1]: https://github.com/reactjs/react-router/compare/v2.8.0...v2.8.1
+[#3808]: https://github.com/reactjs/react-router/pull/3808
+
+
+## [v2.8.0]
+> Sep 9, 2016
+
+- **Feature:** Support omitting `to` on `<Link>` ([#3803])
+- **Refactor:** Use `history.replace` instead of `history.transitionTo` for redirects ([#3799])
+
+[v2.8.0]: https://github.com/reactjs/react-router/compare/v2.7.0...v2.8.0
+[#3799]: https://github.com/reactjs/react-router/pull/3799
+[#3803]: https://github.com/reactjs/react-router/pull/3803
+
+
+## [v2.7.0]
+> Aug 20, 2016
+
+- **Feature:** Support `router` as a prop on `withRouter`-wrapped components for overriding the router object from context ([#3729])
+- **Feature:** Add `withRef` option to `withRouter` that enables `getWrappedInstance` ([#3735], [#3740])
+- **Bugfix:** Warn on invalid router middlewares ([#3717])
+
+[v2.7.0]: https://github.com/reactjs/react-router/compare/v2.6.1...v2.7.0
+[#3717]: https://github.com/reactjs/react-router/pull/3717
+[#3729]: https://github.com/reactjs/react-router/pull/3729
+[#3735]: https://github.com/reactjs/react-router/pull/3735
+[#3740]: https://github.com/reactjs/react-router/pull/3740
+
+
 ## [v2.6.1]
 > Jul 29, 2016